requirement.txt
certifi==2021.10.8
charset-normalizer==2.0.12
facebook-sdk==3.1.0
idna==3.3
importlib-metadata==4.11.1
Pillow==9.0.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.3
qrcode==7.3.1
requests==2.27.1
six==1.16.0
speaklater==1.3
urllib3==1.26.8
zipp==3.7.0
